refactor: reuse actor related util functions
